dhi-kubernetes-cluster-autoscaler/requirements.txt
# DHI Kubernetes Cluster Autoscaler Dependencies
kubernetes>=28.1.0
docker>=7.0.0
aiohttp>=3.9.0
asyncio>=3.4.3
pyyaml>=6.0.1
prometheus-client>=0.19.0
cryptography>=41.0.0
aiofiles>=23.2.1
python-json-logger>=2.0.7
httpx>=0.26.0
uvicorn>=0.25.0
fastapi>=0.109.0
pydantic>=2.5.0
redis>=5.0.0
psutil>=5.9.0
web3>=6.11.0
eth-account>=0.10.0
